backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
requests
pytesseract
Pillow
numpy
opencv-python-headless
pdf2image
